2022/06/20

+++ Assets/TowerDataSc.cs
@@ -8,12 +8,12 @@ public class TowerData
     public string Name;
     public int TowerStep;
     public int TowerCode;
-    public int Damage;
+    public float Damage;
     public float Delay;
     public float Range;
-    public int Critical;
+    public float Critical;
     public float UpgradeAtk;
-    public int UpgradeCri;
+    public float UpgradeCri;
     public int TowerPrice;
     public int UpgradePrice;
 }
